Refactor docker in integration test
- create common docker utils used by different platforms
- remove docker from platform initialized; now it's used in PostRun
