+++ integration/nwo/weaver/platform.go
@@ -19,19 +19,17 @@ import (
 	"text/template"
 	"time"
 
-	docker "github.com/fsouza/go-dockerclient"
-	. "github.com/onsi/gomega"
-	"github.com/tedsuo/ifrit/grouper"
-
 	api2 "github.com/hyperledger-labs/fabric-smart-client/integration/nwo/api"
 	"github.com/hyperledger-labs/fabric-smart-client/integration/nwo/common"
+	"github.com/hyperledger-labs/fabric-smart-client/integration/nwo/common/docker"
 	fabric2 "github.com/hyperledger-labs/fabric-smart-client/integration/nwo/fabric"
-	"github.com/hyperledger-labs/fabric-smart-client/integration/nwo/fabric/helpers"
 	"github.com/hyperledger-labs/fabric-smart-client/integration/nwo/fabric/network"
 	"github.com/hyperledger-labs/fabric-smart-client/integration/nwo/fabric/topology"
 	"github.com/hyperledger-labs/fabric-smart-client/integration/nwo/fabric/topology/fabric"
 	"github.com/hyperledger-labs/fabric-smart-client/integration/nwo/fsc"
 	"github.com/hyperledger-labs/fabric-smart-client/integration/nwo/weaver/interop"
+	. "github.com/onsi/gomega"
+	"github.com/tedsuo/ifrit/grouper"
 )
 
 const (
@@ -85,33 +83,18 @@ type Platform struct {
 	Builder           api2.Builder
 	EventuallyTimeout time.Duration
 
-	NetworkID    string
-	DockerClient *docker.Client
+	NetworkID string
 
 	colorIndex int
 }
 
 func NewPlatform(ctx api2.Context, t api2.Topology, builder api2.Builder) *Platform {
-	helpers.AssertImagesExist(RequiredImages...)
-
-	dockerClient, err := docker.NewClientFromEnv()
-	Expect(err).NotTo(HaveOccurred())
-	networkID := common.UniqueName()
-	_, err = dockerClient.CreateNetwork(
-		docker.CreateNetworkOptions{
-			Name:   networkID,
-			Driver: "bridge",
-		},
-	)
-	Expect(err).NotTo(HaveOccurred())
-
 	return &Platform{
 		Context:           ctx,
 		Topology:          t.(*Topology),
 		Builder:           builder,
 		EventuallyTimeout: 10 * time.Minute,
-		NetworkID:         networkID,
-		DockerClient:      dockerClient,
+		NetworkID:         common.UniqueName(),
 	}
 }
 
@@ -150,6 +133,16 @@ func (p *Platform) Members() []grouper.Member {
 }
 
 func (p *Platform) PostRun(bool) {
+	// getting our docker helper, check required images exists and launch a docker network
+	d, err := docker.GetInstance()
+	Expect(err).NotTo(HaveOccurred())
+
+	err = d.CheckImagesExist(RequiredImages...)
+	Expect(err).NotTo(HaveOccurred())
+
+	err = d.CreateNetwork(p.NetworkID)
+	Expect(err).NotTo(HaveOccurred())
+
 	for _, relay := range p.Topology.Relays {
 		cc, err := p.PrepareInteropChaincode(relay)
 		Expect(err).NotTo(HaveOccurred())
@@ -208,47 +201,18 @@ func (p *Platform) PostRun(bool) {
 }
 
 func (p *Platform) Cleanup() {
-	if p.DockerClient == nil {
-		return
-	}
+	//if p.DockerClient == nil {
+	//	return
+	//}
 
 	cleanupFunc()
 
-	nw, err := p.DockerClient.NetworkInfo(p.NetworkID)
-	if _, ok := err.(*docker.NoSuchNetwork); err != nil && ok {
-		return
-	}
+	d, err := docker.GetInstance()
 	Expect(err).NotTo(HaveOccurred())
 
-	containers, err := p.DockerClient.ListContainers(docker.ListContainersOptions{All: true})
-	Expect(err).NotTo(HaveOccurred())
-	for _, c := range containers {
-		for _, name := range c.Names {
-			p.DockerClient.DisconnectNetwork(p.NetworkID, docker.NetworkConnectionOptions{
-				Force:     true,
-				Container: c.ID,
-			})
-			if strings.HasPrefix(name, "/driver") || strings.HasPrefix(name, "/relay") {
-				err := p.DockerClient.RemoveContainer(docker.RemoveContainerOptions{ID: c.ID, Force: true})
-				Expect(err).NotTo(HaveOccurred())
-				break
-			}
-		}
-	}
-
-	images, err := p.DockerClient.ListImages(docker.ListImagesOptions{All: true})
-	Expect(err).NotTo(HaveOccurred())
-	for _, i := range images {
-		for _, tag := range i.RepoTags {
-			if strings.HasPrefix(tag, p.NetworkID) {
-				err := p.DockerClient.RemoveImage(i.ID)
-				Expect(err).NotTo(HaveOccurred())
-				break
-			}
-		}
-	}
-
-	err = p.DockerClient.RemoveNetwork(nw.ID)
+	err = d.Cleanup(p.NetworkID, func(name string) bool {
+		return strings.HasPrefix(name, "/driver") || strings.HasPrefix(name, "/relay")
+	})
 	Expect(err).NotTo(HaveOccurred())
 }
